♻️ Refactor: 학습자의 말하기연습 화면 폴더 구조 변경 #1
