+++ KoApp-Android/app/src/main/java/com/dagachi/koapp_android/view/main/learner/speaking/adapter/SpeakingSubjectAdapter.kt
@@ -1,9 +1,10 @@
-package com.dagachi.koapp_android.view.main.learner.speaking
+package com.dagachi.koapp_android.view.main.learner.speaking.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.dagachi.koapp_android.databinding.ItemSpeakingSubjectBinding
+import com.dagachi.koapp_android.view.main.learner.speaking.SpeakingSubject
 
 /* 학습자의 말하기 연습 주제별 화면의 RV 어댑터 */
 class SpeakingSubjectAdapter(private val subjectList: ArrayList<SpeakingSubject>): RecyclerView.Adapter<SpeakingSubjectListViewHolder>() {
